+++ YOY.UserAPI/Models/v1/Miscellaneous/Location/POCO/ProcessedLocation.cs
@@ -8,7 +8,7 @@ namespace YOY.UserAPI.Models.v1.Miscellaneous.Location.POCO
     public class ProcessedLocation
     {
         public bool ValidLocation { set; get; }
-        public double? Latitude { set; get; }
-        public double? Longitude { set; get; }
+        public decimal? Latitude { set; get; }
+        public decimal? Longitude { set; get; }
     }
 }
